profile page with both modals, which close upon submitting
